removes the set operation on other_labels
When the messages comes in from the server it
is atomic to the label that was set and not a bulk
operation. Depending on the the order received of
messages that fired in rapid order, you could have a
situation where you added two labels. If the messages
are out of order it will override the array with the
wrong state.